requirements.txt
altgraph==0.16.1
attrs==19.1.0
autopep8==1.4.3
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
bs4==0.0.1
certifi==2018.8.24
chardet==3.0.4
Click==7.0
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
Django==2.1.3
django-mssql==1.8
EasyProcess==0.2.5
entrypoints==0.3
et-xmlfile==1.0.1
Flask==1.1.1
future==0.17.1
graphviz==0.10.1
idna==2.7
ipykernel==5.1.1
ipython==7.6.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.15
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.14.0
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
macholib==1.11
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.1
openpyxl==2.6.2
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
pefile==2019.4.18
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
pycodestyle==2.5.0
pyftpdlib==1.5.5
Pygments==2.4.2
PyInstaller==3.4
pyperclip==1.7.0
pyrsistent==0.15.2
pyscreenshot==0.5.1
python-dateutil==2.8.0
pytz==2018.7
pywin32==224
pywin32-ctypes==0.2.0
pywinpty==0.5.5
pyzmq==18.0.2
qtconsole==4.5.1
requests==2.19.1
requests-cache==0.4.13
selenium==3.14.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.8
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.23
virtualenv==16.0.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.4.2
xlrd==1.2.0
